+++ search-films/src/components/Feedback/Feedback.props.ts
@@ -1,0 +1,8 @@
+import { FilmReviewApi } from '../../api/film.request';
+
+export interface FeedbackProps {
+    data: {
+        description: string;
+        review: FilmReviewApi
+    }
+};